Renamed to dimensional.
