+++ src/Numeric/Units/Dimensional/Dimensions.hs
@@ -14,12 +14,12 @@ arithmetic operators over term level dimensions and aliases for the base term-le
 to avoid namespace pollution. These features are available directly from
 "Numeric.Units.Dimensional.DK.Dimensions.TermLevel" if desired.
 -}
-module Numeric.Units.Dimensional.DK.Dimensions
+module Numeric.Units.Dimensional.Dimensions
 (
-  module Numeric.Units.Dimensional.DK.Dimensions.TermLevel,
-  module Numeric.Units.Dimensional.DK.Dimensions.TypeLevel
+  module Numeric.Units.Dimensional.Dimensions.TermLevel,
+  module Numeric.Units.Dimensional.Dimensions.TypeLevel
 )
 where
 
-import Numeric.Units.Dimensional.DK.Dimensions.TermLevel hiding ((*), (/), (^), recip, dLength, dMass, dTime, dElectricCurrent, dThermodynamicTemperature, dAmountOfSubstance, dLuminousIntensity)
-import Numeric.Units.Dimensional.DK.Dimensions.TypeLevel
+import Numeric.Units.Dimensional.Dimensions.TermLevel hiding ((*), (/), (^), recip, dLength, dMass, dTime, dElectricCurrent, dThermodynamicTemperature, dAmountOfSubstance, dLuminousIntensity)
+import Numeric.Units.Dimensional.Dimensions.TypeLevel
