+++ tests/Test.hs
@@ -1,11 +1,11 @@
-import qualified Numeric.Units.Dimensional.DK.Test
-import qualified Numeric.Units.Dimensional.DK.QuantitiesTest
+import qualified Numeric.Units.Dimensional.Test
+import qualified Numeric.Units.Dimensional.QuantitiesTest
 import System.Exit
 
 main :: IO ()
 main = do
-         Numeric.Units.Dimensional.DK.QuantitiesTest.main
-         ok <- Numeric.Units.Dimensional.DK.Test.main
+         Numeric.Units.Dimensional.QuantitiesTest.main
+         ok <- Numeric.Units.Dimensional.Test.main
          if ok
            then exitSuccess
            else exitFailure
